requirements.txt
aiogram==3.1.1
SQLAlchemy==2.0.23
openai==1.3.5
apscheduler==3.10.4
aiosqlite
python-dotenv>=1.0.0
